apply feedback; fix tests; extract saga into own file
