+++ frontend/javascripts/oxalis/model/sagas/root_saga.js
@@ -17,7 +17,8 @@ import {
   warnAboutSegmentationOpacity,
   watchAnnotationAsync,
 } from "oxalis/model/sagas/annotation_saga";
-import { watchDataRelevantChanges, prefetchHistorgramData } from "oxalis/model/sagas/prefetch_saga";
+import { watchDataRelevantChanges } from "oxalis/model/sagas/prefetch_saga";
+import { loadHistorgramData } from "oxalis/model/sagas/load_histogram_data";
 import { watchSkeletonTracingAsync } from "oxalis/model/sagas/skeletontracing_saga";
 import ErrorHandling from "libs/error_handling";
 import handleMeshChanges from "oxalis/model/sagas/handle_mesh_changes";
@@ -48,7 +49,7 @@ function* restartableSaga(): Saga<void> {
       _call(watchVolumeTracingAsync),
       _call(watchAnnotationAsync),
       _call(watchDataRelevantChanges),
-      _call(prefetchHistorgramData),
+      _call(loadHistorgramData),
       _call(isosurfaceSaga),
       _call(watchTasksAsync),
       _call(handleMeshChanges),
